+++ qubes/tests/__init__.py
@@ -1439,6 +1439,7 @@ def load_tests(loader, tests, pattern):  # pylint: disable=unused-argument
             'qubes.tests.integ.grub',
             'qubes.tests.integ.devices_block',
             'qubes.tests.integ.devices_pci',
+            'qubes.tests.integ.qrexec',
             'qubes.tests.integ.dom0_update',
             'qubes.tests.integ.vm_update',
             'qubes.tests.integ.network',
